Offset into the file can be size_t and add some casts to remove warnings.

+++ usr.bin/file/magic.h
@@ -1,4 +1,4 @@
-/* $OpenBSD: magic.h,v 1.8 2015/08/11 22:12:48 nicm Exp $ */
+/* $OpenBSD: magic.h,v 1.9 2015/10/02 18:06:27 deraadt Exp $ */
 
 /*
  * Copyright (c) 2015 Nicholas Marriott <nicm@openbsd.org>
@@ -160,7 +160,7 @@ struct magic_state {
 
 	const char		*base;
 	size_t			 size;
-	int64_t			 offset;
+	size_t			 offset;
 };
 
 #define MAGIC_TEST_TEXT 0x1
